+++ docs/release-notes.md
@@ -16,7 +16,7 @@ SPDX-License-Identifier: CC-BY-4.0
 
 ### Enhancements
 
-- Add `include_objective_constant` parameter to `n.optimize()` and `n.optimize.create_model()` to control whether the objective constant (capital costs of existing infrastructure) is included as a variable in the objective function and assigned to the network. Setting to `False` sets `n.objective_constant` to zero and improves LP numerical conditioning. The default will change from `True` to `False` in a future version. (<!-- md:pr 503 -->)
+- Add `include_objective_constant` parameter to `n.optimize()` and `n.optimize.create_model()` to control whether the objective constant (capital costs of existing infrastructure) is included as a variable in the objective function and assigned to the network. Setting to `False` sets `n.objective_constant` to zero and improves LP numerical conditioning. The default will change from `True` to `False` in version 2.0. Can also be configured via `pypsa.options.params.optimize.include_objective_constant`. (<!-- md:pr 1509 -->)
 
 ## [**v1.0.6**](https://github.com/PyPSA/PyPSA/releases/tag/v1.0.6) <small>22nd December 2025</small> { id="v1.0.6" }